% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasstationqueue.R
\name{get_avg_queue_length}
\alias{get_avg_queue_length}
\title{Calculate average queue length}
\usage{
get_avg_queue_length(simulation_results)
}
\arguments{
\item{simulation_results}{Results from simulate_gas_station function}
}
\value{
Average queue length
}
\description{
Calculate average queue length
}
