% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gasstationqueue.R
\name{get_avg_waiting_time}
\alias{get_avg_waiting_time}
\title{Calculate average waiting time}
\usage{
get_avg_waiting_time(simulation_results)
}
\arguments{
\item{simulation_results}{Results from simulate_gas_station function}
}
\value{
Average waiting time in minutes
}
\description{
Calculate average waiting time
}
